# Conflicts:
#	frontend/src/components/signup/WriteInfo.vue

+++ backend/spring/src/main/java/com/ssafy/db/entity/Book_review.java
@@ -20,7 +20,7 @@ public class Book_review extends BaseEntity{
 	@JoinColumn(name = "book_bid")
 	private Book book;
 	
-	private Double point;
+	private Integer point;
 	private String content;
 
 }
